@@ -1,6 +1,6 @@
 ﻿using CleanTemplate.Common.Utilities;
 using Microsoft.AspNetCore.Mvc.Controllers;
-using Microsoft.OpenApi.Models;
+using Microsoft.OpenApi;
 using Pluralize.NET;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
