@@ -1,9 +1,9 @@
 ﻿namespace CleanTemplate.ApiFramework.Tools
 {
-    using Microsoft.AspNetCore.Mvc.ApiExplorer;
+    using Asp.Versioning.ApiExplorer;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Options;
-    using Microsoft.OpenApi.Models;
+    using Microsoft.OpenApi;
     using Swashbuckle.AspNetCore.SwaggerGen;
     using System;
     public class ConfigureSwaggerOptions : IConfigureOptions<SwaggerGenOptions>
